refactor, anable hot reload, options for naming
